# docs/openapi.yaml
openapi: 3.1.0
info:
  title: QK 教培 CRM 管理系统 API
  version: 1.1.12
  summary: 线索、商机、客户全流程管理接口
  description: |
    面向教培行业销售团队的后台管理接口，覆盖线索录入、分配、跟进、转商机、再分配、
    再跟进、转客户的全流程，以及部门、角色、用户、课程、活动等基础数据维护。

    ## 通用约定

    ### 统一响应结构
    所有接口返回 `Result` 对象：`code` 为 `1` 表示成功、`0` 表示失败。
    **业务失败同样返回 HTTP 200**，由调用方依据 `code` 判断；`msg` 为可直接展示给用户的中文提示。
    无业务数据时 `data` 为 `null`，且该字段会**整个省略**（服务端 `Result` 标注 `@JsonInclude(NON_NULL)`），
    判断成功请用 `code == 1`，不要依赖 `data` 字段是否存在。

    ### 认证
    除 `POST /login` 外，所有接口都要求在请求头 `token` 中携带登录接口签发的 JWT 令牌。
    令牌缺失、被篡改、已过期，或令牌对应的账号已被停用 / 删除时，接口直接返回 **HTTP 401 且响应体为空**，
    不返回 `Result` 结构。
    令牌默认有效期 24 小时。

    ### 授权（权限点模型）
    已登录不等于有权限。每个接口在代码里声明自己需要的**权限点**（如 `clue:track`、`user:delete`，
    完整目录见 `GET /permissions`），账号能不能调用取决于「它的角色被授予了哪些权限点」。
    角色是数据，可以自由增删改，**改角色名/标识不影响任何权限**；授权关系存在 `role_permission` 表，
    用 `GET /roles/{id}/permissions` 查看、`PUT /roles/{id}/permissions` 覆盖式配置（改完立即生效）。
    未声明权限的接口对所有人拒绝（默认拒绝），唯一例外是 `POST /login`。
    内置超级管理员角色（`role.is_super = 1`）天然拥有全部权限，不可删除、也不支持单独调整权限。
    登录响应里的 `permissions` 即当前账号的权限码列表，前端用它控制菜单与按钮（不要用 `roleLabel`）。

    ### 分页
    分页接口的查询参数为 `page`（页码，从 1 开始，默认 1）与 `pageSize`（每页条数，默认 10，**单页最大 200**），
    响应数据统一为 `{ "total": 总记录数, "rows": [...] }`。
    `page` / `pageSize` 必须为正整数：传空值、0、负数或超过 200 时返回 `code = 0`，`msg` 为对应的字段提示。

    ### 错误语义（HTTP 状态码）
    | 状态码 | 触发场景 | 响应体 |
    | --- | --- | --- |
    | `200` | 成功；以及**业务失败**（数据不存在、唯一键冲突、文件类型不合法、状态不允许等） | `Result`，失败时 `code = 0`，按 `msg` 提示 |
...

...

...

...

...

...

...

...
servers:
...
security:
  - token: []
...
paths:
...
  /activities:
    get:
      tags: [活动管理]
      summary: 条件分页查询活动
      operationId: listActivities
      parameters:
        - name: channel
          in: query
          description: 渠道来源
          schema: { $ref: '#/components/schemas/Channel' }
        - name: type
          in: query
          description: 活动类型
          schema: { $ref: '#/components/schemas/ActivityType' }
        - name: activityStatus
          in: query
          description: 活动状态，1 未开始 / 2 进行中 / 3 已结束；不传表示不筛选
          schema: { $ref: '#/components/schemas/ActivityStatus' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 分页结果，按更新时间倒序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultActivityPage'
    post:
      tags: [活动管理]
      summary: 新增活动
      description: 该接口会写入一条操作日志。
      operationId: addActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityBody'
            example:
              channel: 1
              name: 暑期 Java 就业班折扣
              startTime: '2025-06-01 00:00:00'
              endTime: '2025-06-30 23:59:59'
              description: 暑期报名享 8 折
              type: 1
              discount: 8.0
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
    put:
      tags: [活动管理]
      summary: 修改活动
      description: 该接口会写入一条操作日志。
      operationId: updateActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityBody'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /activities/type/{type}:
    get:
      tags: [活动管理]
      summary: 根据活动类型查询活动
      operationId: listActivitiesByType
      parameters:
        - name: type
          in: path
          required: true
          description: 活动类型
          schema: { $ref: '#/components/schemas/ActivityType' }
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 活动列表，无匹配时返回空数组
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultActivityList'
  /activities/{id}:
    get:
      tags: [活动管理]
      summary: 根据ID查询活动
      operationId: getActivityById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 活动信息；不存在时 `code = 0`、`msg = 活动不存在`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultActivity'
    delete:
      tags: [活动管理]
      summary: 删除活动
      description: 该接口会写入一条操作日志。
      operationId: deleteActivity
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
...
  /businesses:
    get:
      tags: [商机管理]
      summary: 商机列表查询
      description: |
        默认只返回仍在跟进的商机：`status` 为 `4 回收` 与 `5 转客户` 的商机不出现在此列表。
        显式传 `status` 时按其取值筛选（前端状态下拉里包含这两个状态，选中后应能看到对应数据）。
        按修改时间倒序。
      operationId: listBusinesses
      parameters:
        - name: businessId
          in: query
          description: 商机ID，精确匹配
          schema: { type: integer, format: int64 }
        - name: name
          in: query
          description: 客户姓名，模糊匹配
          schema: { type: string }
        - name: phone
          in: query
          description: 手机号，模糊匹配
          schema: { type: string }
        - name: status
          in: query
          description: 商机状态；不传时排除 4 回收、5 转客户
          schema: { $ref: '#/components/schemas/BusinessStatus' }
        - name: assignName
          in: query
          description: 归属人姓名，模糊匹配
          schema: { type: string }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 分页结果，`rows` 中 `trackRecords`、`trackStatus`、`keyItems`、`record` 为 `null`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultBusinessPage'
    post:
      tags: [商机管理]
      summary: 新增商机
      description: |
        新增的商机一律置为 `1 待分配` 且归属人清空，请求体中的 `status`、`userId` 不生效。
        手机号唯一，重复时返回 `msg = 该手机号已录入商机`。该接口会写入一条操作日志。
      operationId: addBusiness
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessBody'
            example:
              phone: '15300000001'
              channel: 2
              name: 承娟
              gender: 2
              age: 19
              wechat: cj2839
              qq: '2595964758'
              subject: 1
              degree: 4
              jobStatus: 1
              courseId: 1
              remark: 无
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
    put:
      tags: [商机管理]
      summary: 跟进商机
      description: |
        将商机状态置为 `3 跟进中`，并新增一条商机跟进记录。
        请求体中除 `id` 外的字段为可选，未传字段保持原值。
        `keyItems` 为沟通重点数组，落库时序列化为字符串。该接口会写入一条操作日志。
      operationId: trackBusiness
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessTrackBody'
            example:
              id: 21
              trackStatus: 1
              keyItems: [课程, 时间]
              nextTime: '2025-06-23 10:00:00'
              record: 了解了课程及上课时间
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /businesses/pool:
    get:
      tags: [商机管理]
      summary: 公海池列表查询
      description: 只返回被踢回公海（`status = 4`）的商机。按修改时间倒序。
      operationId: listPoolBusinesses
      parameters:
        - name: businessId
          in: query
          description: 商机ID，精确匹配
          schema: { type: integer, format: int64 }
        - name: phone
          in: query
          description: 手机号，模糊匹配
          schema: { type: string }
        - name: name
          in: query
          description: 客户姓名，模糊匹配
          schema: { type: string }
        - name: subject
          in: query
          description: 意向学科
          schema: { $ref: '#/components/schemas/Subject' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultBusinessPage'
  /businesses/assign/{businessId}/{userId}:
    put:
      tags: [商机管理]
      summary: 分配商机给指定用户
      description: 将商机归属人设为指定用户并置为 `2 待跟进`。该接口会写入一条操作日志。
      operationId: assignBusiness
      parameters:
        - name: businessId
          in: path
          required: true
          description: 商机ID
          schema: { type: integer, format: int64 }
        - name: userId
          in: path
          required: true
          description: 被分配的用户ID
          schema: { type: integer, format: int64 }
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /businesses/back/{id}:
    put:
      tags: [商机管理]
      summary: 将商机踢回公海
      description: 将商机置为 `4 回收` 并清空归属人。该接口会写入一条操作日志。
      operationId: backBusinessToPool
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /businesses/toCustomer/{id}:
    post:
      tags: [商机管理]
      summary: 将商机转为客户
      description: |
        将商机置为 `5 转客户`，并按商机信息复制生成一条客户，同时记录来源商机ID。
        该接口会写入一条操作日志。
      operationId: convertBusinessToCustomer
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /businesses/{id}:
    get:
      tags: [商机管理]
      summary: 根据ID查询商机详细信息
      operationId: getBusinessById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: |
            商机详细信息，`trackRecords` 为该商机的全部跟进记录（按创建时间倒序），
            每条记录带出跟进人姓名；`courseName` 为意向课程名称。
            商机不存在时 `code = 0`、`msg = 商机不存在`。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultBusiness'
...
components:
  securitySchemes:
    token:
      type: apiKey
      in: header
      name: token
      description: |
        登录接口返回的 JWT 令牌。缺失、被篡改或已过期时接口返回 HTTP 401 且响应体为空。
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: 主键ID，64 位整型（对应数据库 bigint unsigned）
      schema: { type: integer, format: int64 }
...
    Page:
      name: page
      in: query
      description: 页码，从 1 开始
      schema: { type: integer, format: int32, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      description: 每页记录数
      schema: { type: integer, format: int32, minimum: 1, default: 10 }
  responses:
    Ok:
      description: 操作成功，`code = 1`、`data` 为 `null`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
    Unauthorized:
      description: 令牌缺失、被篡改或已过期。响应体为空，不返回 `Result` 结构。
...
  schemas:
...
    Result:
      type: object
      description: 统一响应结构。业务失败同样返回 HTTP 200，由 `code` 区分。
      properties:
        code:
          type: integer
          description: 1:成功, 0:失败
          enum: [1, 0]
        msg:
          type: string
          description: 提示信息，成功时固定为 `success`
        data:
          description: 业务数据，无数据时为 `null`
      required: [code, msg]
    DateTime:
      type: string
      description: 时间，输出格式 `yyyy-MM-dd HH:mm:ss`
      example: '2025-06-01 00:00:00'

    # ---------- 枚举 ----------
    Channel:
      type: integer
      format: int32
      description: 渠道来源，1:线上活动, 2:推广介绍
      enum: [1, 2]
    Subject:
      type: integer
      format: int32
      description: 意向学科，1:AI智能应用开发(Java), 2:AI大模型开发(Python), 3:AI鸿蒙开发, 4:AI大数据, 5:AI嵌入式, 6:AI测试, 7:AI运维
      enum: [1, 2, 3, 4, 5, 6, 7]
    Gender:
      type: integer
      format: int32
      description: 性别，1:男, 2:女
      enum: [1, 2]
    Degree:
      type: integer
      format: int32
      description: 学历，1:高中, 2:中专, 3:大专, 4:本科, 5:硕士, 6:博士, 7:其他
      enum: [1, 2, 3, 4, 5, 6, 7]
    JobStatus:
      type: integer
      format: int32
      description: 在职情况，1:在职, 0:离职
      enum: [0, 1]
...
    BusinessStatus:
      type: integer
      format: int32
      description: 商机状态，1:待分配, 2:待跟进, 3:跟进中, 4:回收, 5:转客户
      enum: [1, 2, 3, 4, 5]
    BusinessTrackStatus:
      type: integer
      format: int32
      description: 跟进状态，1:接通, 2:拒绝, 3:无人接听
      enum: [1, 2, 3]
    ActivityType:
      type: integer
      format: int32
      description: 活动类型，1:课程折扣, 2:代金券
      enum: [1, 2]
    ActivityStatus:
      type: integer
      format: int32
      description: |
        活动状态（查询条件，不落库，由 `startTime` / `endTime` 与当前时间比较得出）：
        1 未开始（startTime > now）、2 进行中（startTime <= now <= endTime）、3 已结束（endTime < now）
      enum: [1, 2, 3]
...
    ActivityVO:
      type: object
      properties:
        id: { type: integer, format: int64, description: 活动ID }
        channel: { $ref: '#/components/schemas/Channel' }
        name: { type: string, description: 活动名称 }
        startTime: { $ref: '#/components/schemas/DateTime' }
        endTime: { $ref: '#/components/schemas/DateTime' }
        description:
          type: [string, 'null']
          description: 活动简介
        type: { $ref: '#/components/schemas/ActivityType' }
        discount:
          anyOf:
            - type: number
            - type: 'null'
          description: 课程折扣，例如 `8.0` 表示 8 折。与 `voucher` 二选一
          example: 8.0
        voucher:
          anyOf:
            - type: integer
              format: int32
            - type: 'null'
          description: 课程代金券，单位：元。与 `discount` 二选一
        createTime: { $ref: '#/components/schemas/DateTime' }
        updateTime: { $ref: '#/components/schemas/DateTime' }
    ActivityBody:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 活动ID，修改时必填
        channel: { $ref: '#/components/schemas/Channel' }
        name: { type: string, description: 活动名称 }
        startTime: { $ref: '#/components/schemas/DateTime' }
        endTime: { $ref: '#/components/schemas/DateTime' }
        description:
          type: [string, 'null']
          description: 活动简介
        type: { $ref: '#/components/schemas/ActivityType' }
        discount:
          anyOf:
            - type: number
            - type: 'null'
          description: 课程折扣。`type = 1` 时填写
        voucher:
          anyOf:
            - type: integer
              format: int32
            - type: 'null'
          description: 课程代金券，单位：元。`type = 2` 时填写
      required: [channel, name, startTime, endTime, type]

    # ---------- 线索 ----------
...
    Business:
      type: object
      properties:
        id: { type: integer, format: int64, description: 商机ID }
        name:
          type: [string, 'null']
          description: 客户姓名
        phone: { type: string, description: 手机号，唯一 }
        gender:
          anyOf:
            - $ref: '#/components/schemas/Gender'
            - type: 'null'
        age:
          anyOf:
            - type: integer
              format: int32
            - type: 'null'
          description: 年龄
        wechat:
          type: [string, 'null']
          description: 微信号
        qq:
          type: [string, 'null']
          description: QQ号
        subject:
          anyOf:
            - $ref: '#/components/schemas/Subject'
            - type: 'null'
        courseId:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: 意向课程ID
        degree:
          anyOf:
            - $ref: '#/components/schemas/Degree'
            - type: 'null'
        jobStatus:
          anyOf:
            - $ref: '#/components/schemas/JobStatus'
            - type: 'null'
        channel: { $ref: '#/components/schemas/Channel' }
        remark:
          type: [string, 'null']
          description: 备注
        status: { $ref: '#/components/schemas/BusinessStatus' }
        userId:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: 归属人ID
        clueId:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: 来源线索ID
        nextTime:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
          description: 下次跟进时间
        createTime: { $ref: '#/components/schemas/DateTime' }
        updateTime: { $ref: '#/components/schemas/DateTime' }
        assignName:
          type: [string, 'null']
          description: 归属人姓名
        courseName:
          type: [string, 'null']
          description: 意向课程名称，仅商机详情接口返回
        trackRecords:
          anyOf:
            - type: array
              items: { $ref: '#/components/schemas/BusinessTrackRecord' }
            - type: 'null'
          description: 跟进记录列表，仅商机详情接口返回
        trackStatus:
          anyOf:
            - $ref: '#/components/schemas/BusinessTrackStatus'
            - type: 'null'
          description: 本次跟进状态，仅用于接收 `PUT /businesses` 的请求参数，查询接口中为 `null`
        keyItems:
          anyOf:
            - type: array
              items: { type: string }
            - type: 'null'
          description: 本次沟通重点，仅用于接收 `PUT /businesses` 的请求参数，查询接口中为 `null`
        record:
          type: [string, 'null']
          description: 本次沟通纪要，仅用于接收 `PUT /businesses` 的请求参数，查询接口中为 `null`
    BusinessTrackRecord:
      type: object
      description: 商机跟进记录
      properties:
        id: { type: integer, format: int64, description: 跟进记录ID }
        businessId: { type: integer, format: int64, description: 商机ID }
        userId: { type: integer, format: int64, description: 跟进人ID }
        trackStatus: { $ref: '#/components/schemas/BusinessTrackStatus' }
        keyItems:
          type: [string, 'null']
          description: 沟通重点，落库时由数组序列化为字符串，例如 `[课程, 时间]`
        nextTime:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        record:
          type: [string, 'null']
          description: 沟通纪要
        createTime: { $ref: '#/components/schemas/DateTime' }
        assignName:
          type: [string, 'null']
          description: 跟进人姓名
    BusinessBody:
      type: object
      description: 新增商机的请求体，`status` 与 `userId` 由服务端强制设置，传入不生效
      properties:
        name:
          type: [string, 'null']
          description: 客户姓名
        phone: { type: string, description: 手机号，唯一 }
        gender:
          anyOf:
            - $ref: '#/components/schemas/Gender'
            - type: 'null'
        age:
          anyOf:
            - type: integer
              format: int32
            - type: 'null'
          description: 年龄
        wechat:
          type: [string, 'null']
          description: 微信号
        qq:
          type: [string, 'null']
          description: QQ号
        subject:
          anyOf:
            - $ref: '#/components/schemas/Subject'
            - type: 'null'
        courseId:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: 意向课程ID
        degree:
          anyOf:
            - $ref: '#/components/schemas/Degree'
            - type: 'null'
        jobStatus:
          anyOf:
            - $ref: '#/components/schemas/JobStatus'
            - type: 'null'
        channel:
          anyOf:
            - $ref: '#/components/schemas/Channel'
            - type: 'null'
        remark:
          type: [string, 'null']
          description: 备注
      # channel 按页面原型（2.11 渠道来源，选填）与接口文档均为非必须
      required: [phone]
    BusinessTrackBody:
      type: object
      description: 跟进商机的请求体，未传字段保持原值
      properties:
        id: { type: integer, format: int64, description: 商机ID }
        trackStatus: { $ref: '#/components/schemas/BusinessTrackStatus' }
        keyItems:
          anyOf:
            - type: array
              items: { type: string }
            - type: 'null'
          description: 沟通重点
        nextTime:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
          description: 下次跟进时间
        record:
          type: [string, 'null']
          description: 沟通纪要
      required: [id]

    # ---------- 客户 ----------
...
    ActivityPage:
      type: object
      properties:
        total: { type: integer, format: int64, description: 总记录数 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/ActivityVO' }
      required: [total, rows]
...
    BusinessPage:
      type: object
      properties:
        total: { type: integer, format: int64, description: 总记录数 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/Business' }
      required: [total, rows]
...
    ResultActivityPage:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ActivityPage' }
    ResultActivityList:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ActivityVO' }
    ResultActivity:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              anyOf:
                - $ref: '#/components/schemas/ActivityVO'
                - type: 'null'
...
    ResultBusinessPage:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/BusinessPage' }
    ResultBusiness:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              anyOf:
                - $ref: '#/components/schemas/Business'
                - type: 'null'
...